time: Solaris fix: don't assign to env_list.
* time.c (env_list): The env_list variable is static in
lib.c. The idea here was that since we modified the
environment with setenv, we should invalidate the env hash
which stores a cached copy of it. However, the setenv library
function that is exposed to programs neglects to do this,
which is a much larger issue. Let's delete the assignment to
env_list and deal with the whole issue later.