gc: fix unnecessary full gc request in finalization.
* gc.c (call_finalizers_impl): Objects are only added to
freshobj if they are in the zero generation. We should skip
that entire block of code if the object isn't in that
generation. Not only is it wasteful to execute that code for
the mature generation, but the logic falsely sets the full_gc
flag whenever processing a non-gen-0 object!