gc: add finalization count to objects.
With the finalization count, we don't have to scan the
freshobj array for duplicates when calling finalizers.
However, the limited range of the counter limits how many
times we can register a finalizer against an object.

* gc.c (make_obj): Reset the new fincount field to zero
for a newly minted object.
(call_finalizers_impl): Decrement the fincount for
each object. Only run the freshobj-related logic when
the count hits zero.
(gc_finalize): Bump the fincount of a registered object. If
the counter overflows, throw an exception.

* lib.h (obj_common): Add new field fincount for the
finalization count.

* tree.c (tr_rebuild): Fix up dummy object initializer
to accommodate the new member.

* txr.1: Document that there is a limit on the number
of times an object can be registered for finalization.