gc: streamlining finalization-related code.
* gc.c (prepare_finals): Do not move unreachable objects with
finalizers into generation 0 here.
(call_finalizers_impl): Don't test the object for being in
generation 0; that may not be true since we don't ensure that
in prepare_finals now. Instead check he reachable flag of he
finalization entry; that tells us that we are dealing with an
object that had been found unreachable. When we re-introduce
it into freshobj, we set its generation to zero.  Also, don't
bother with this freshobj logic if the full_gc flag has
been set. That now includes not bothering to set object
generations to zero.