configure: tweak test for inline.
It seems that the Brew people are running into a problem.  The
configure test is detecting that "inline" can be used for
declaring inline functions, but then some functions are not
inlined, requiring an external definition.

On GNU/Linuxes, we detect "static inline", so this doesn't
reproduce.

This is likely triggered by the recent change to compile as
C99 (unless configuring in maintainer mode).

* configure: Let's revise the set of possible definitions of
INLINE that we try. Firstly, we only try "inline" if we are
compiling as C++. If that fails, all the other forms of inline
we try include "static". Let's not mess with anything involving
"extern". Like a bare "inline", it might work in the test
program but cause some sort of failure.