mapcar/maprod: show-stopper bug.
Observed wrong result:

  (mapcar (lambda (. args) (list . args)) '#(1 2 3) '#(4 5 6))
  -> #((1 4) nil nil)

Correct result:

  -> #((1 4) (2 5) (3 6))

This is not specific to vector input; it's broken for
sequences of all types.

Functions affected are mapcar, mapf, mappend, mapdo,
maprod, maprend, maprodo.

* eval.c (map_common, prod_common): Because we reuse the args
structure across iterations, we must reinitialize the fill
and list members. These can be modified by the functionw which
is called. In particular, the arguments are applied, they may
be turned into a list (fill decrements to zero, and a list is
produced).