matcher: support @(and pats ...) operator.
This is implemented using exactly the same code as @(or ...);
the only difference is whether the and or or operator is used
in the expression.

* share/txr/stdlib/match.tl (compiile-or-match): Renamed to
compile-or-parallel match. Some local variables are renamed to
avoid being OR-specific. The operator is extracted from the
pattern, and inserted into the guard expression. That one
insertion is the only differnce between and and or.
(compile-match): Route the or operator to the renamed
function.  Rout the new and operator to it also.