matcher: add optimized special case to hash pattern.
This change causes a key-value pattern like (@a @b)
to be treated specially when @a already has a binding from a
previous pattern. In this case, it behaves like the
trivial key case: the value of @a is looked up to try to find
a single value. If @a is not bound, then the exhaustive
search takes place, using equal equality.

* share/txr/stdlib/match.tl (compile-hash-match): Implement
special case.
(var-pat-p): New function.

* tests/011/patmatch.tl: Existing test case now changes
value. New test case added.

* txr.1: Documented.