printer: do not render @(rcons ...) in .. notation.
This fixes the following print-read consistency issue.
Both of these objects print as @a..@b.

  1> '@(rcons a b)
  @a..b
  2> '(rcons @a b)
  @a..b

We want only the second case. After the this fix:

  1> '(rcons @a b)
  @a..b
  2> '@(rcons a b)
  @(rcons a b)

* lib.c (obj_print_impl): In the sys:expr case, we check
whether the head of the argument is rcons.  If so, we adjust a
few local variables and branch directly to the generic list
case via goto to print the argument as (rcons ...) without
conversion to dotdot range notation.