@(rebind): bugfix: don't clobber right side variable.
* Makefile (tst/tests/000/binding.ok): Pass -B to txr for this
new test.

* match.c (v_rebind): Fix gaping copy-and-paste bug here,
which causes rebind to take on the behavior of local/forget;
it takes all symbols that appear as its arguments from the
environment and produces an environment in which they
don't exist. What we want is to remove the left
variables from the environment, and since that is a nested
pattern, the right way to do that is to flatten it.
Bug reported by Frank Schwidom.

* tests/000/binding.txr: New file.

* tests/000/binding.expected: New file.

* txr.1: Improve documentation of @(rebind), also making
improvements in @(set) documentation.