matcher: new failing backreferencing test case.
* tests/011/patmatch.tl: New test cases that break. The (copy
var-list) logic in the handling of and and or is incomplete.
The bifurcated vars must be merged together into the original
vars. Without this, it looks as if the operator didn't bind
any variables, and they can be repeated again without
backreferencing.  In the broken examples, variable a is taking
on the value 2 instead of mismatching the previous value of 1.