lib: code review of get_special_slot uses.
* lib.c (seq_info): Streamline code. Initialize ret.kind to
SEQ_NOTSEQ to avoid repeating that in multiple else clauses.
Make sure that if the nullify method returns nil, we also
return in that case, and don't call get_special_slot on nil.
(rplaca, rplacd): Change the default case to COBJ, so that the
obj_struct_p inline can be used instead of structp.
Put the default: label on the error case which the struct
check falls through to; NIL now goes there.
(sub): Use obj_struct_p, not structp, in a case where we
already know we have a COBJ.