matcher: don't set flag in last case.
The last case in a match-case or lambda-match does not need to
set the matched flag, since nothing tests it.

* share/txr/stdlib/match.tl (match-case): Rename some local
variables for consistency with lambda-match. Change the
counter from 1, so we can then compare the index of the last
case to the length and avoid emitting the set.
(expand-lambda-match): Same optimization.