matcher: @(as) must always bind fresh variable.
It is documented that as binds a new variable. Furthermore, it
used to be called let. Yet, it back-references. Let's fix it.

* share/txr/stdlib/match.tl (compile-new-var-match): New
function: like compile-var-match but binds new variable, as if
back-referencing didn't exist.
(compile-as-match): Use compile-new-var-match.

* txr.1: Improve as documentation. Clarify that it binds a
fresh variable, and that pattern is processed in its scope.
Include an example with circular structure.