matcher: eliminate use of flags.
* share/txr/stdlib/match.tl (if-match, match-case,
lambda-match): Instead of returning the result from the
case(s), which gets stored in a result variable, and setting a
flag to t, set the result variable inside the case, and return
t. This eliminates the flag. In match-case and lambda-match,
the cases can then be combined into an or form.