compiler: jump skid optimization.
If an (if reg label0) target branches due to reg being nil,
and the target of the branch is another branch instrution
of the form (ifq reg nil label1), we know that that branch is
not taken: the code following that instruction is executed.
THus can jump right to that code.

  (if reg label0)                  (if reg xlabel)
  ...                              ...
  label0                           label0
  (ifq reg nil label1)  -->        (ifq reg nil label1)
  ...                              xlabel
                                   ...

* share/txr/stdlib/optimize.tl (basic-blocks peephole): New
sub-case under (jmp @reg @jlabel).